Allow v4.x & v5.x Operator to Run in a Single Namespace
It is not currently possible to run both the v4.x and v5.x PostgreSQL
Operator in a single namespace due to a naming conflict with the
operators' respectively deployments.

This commit updates the PostgreSQL Operator v5.x installers to name
the v5.x Operator deployment 'pgo' instead of 'postgres-operator' to
allow both versions to run in the same namespace.

Also, to facilitate independent installation and removal of the 4.x
and 5.x Operators, the default service account name has been updated
in 5.x to 'pgo' in place of the current 'postgres-cluster'.