Ensure additional info labels are not in Deployment selectors
Kustomize commonLabels are included even in the match selectors.
This adds some patching so that they are not including in the
match selectors for the PGO Deployment, as match selectors are
immutable and these labels are superfluous.

This adjusts for the behavior added in 4f123b220. If one is
affected by this, one can reinstall the PGO Deployment with

    kubectl delete -k kustomize/install/bases/manager
    kubectl apply -k kustomize/install