Update Examples for the Latest Version of Keycloak
The Keycloak example now works with the latest version of Keycloak.  This includes using
the new env vars and settings that are now available in the latest version of Keycloak.

The PostgresCluster's created for the Keycloak examples (which includes any PostgresCluster's
created for use with Keycloak in the Crunchy Postgres for Kubernetes documentation) have
also been updated to set the "autoCreateUserSchema" annotation to "true".  Without this setting,
Keycloak will crash on startup when attempting to use the public schema.